4 more solution added

+++ 008_String_to_Integer_(atoi).cpp
@@ -39,6 +39,6 @@ public:
                 break;
             }
         }
-        return sum;
+        return (int)sum;
     }
 };
